# BIGTOP-4044: Enhance Bigtop with Concurrent Compilation Support for A…