--- a/packages.gradle
+++ b/packages.gradle
@@ -76,6 +76,27 @@ def safeDelete = { fileName ->
 def getDate() {
   new Date().format('E, dd MMM yyyy HH:mm:ss Z')
 }
+
+def generatePatch(String threadCount, String path) {
+    def patchContent = """
+        diff --git a/.mvn/maven.config b/.mvn/maven.config
+        new file mode 100644
+        index 00000000..e69de29b
+        diff --git a/.mvn/maven.config b/.mvn/maven.config
+        index e69de29b..0bea21c9 100644
+        --- a/.mvn/maven.config
+        +++ b/.mvn/maven.config
+        @@ -0,0 +1 @@
+        +-T${threadCount}
+        \\ No newline at end of file
+    """.stripIndent()
+
+    new File(path, "patch0-maven-parallel-build.diff").with {
+        write(patchContent)
+        println "Patch file created at: ${absolutePath}"
+    }
+}
+
 /**
  * To avoid breaking the compat with existing packages let's use the old style names
  */
@@ -137,6 +158,7 @@ task "bom-json" (description: "List the components of the stack in json format")
         pkg: it.value.pkg,
         relNotes: it.value.relNotes,
         packaging: it.value.packaging,
+        maven_parallel_build: it.value.maven_parallel_build
       ],
       tarball: [
         destination: it.value.tarball.destination,
@@ -394,6 +416,8 @@ def genTasks = { target ->
     delete ("$PKG_BUILD_DIR/deb")
     def final DEB_BLD_DIR = "$PKG_BUILD_DIR/deb/$NAME-${PKG_VERSION}"
     def final DEB_PKG_DIR = "$PKG_BUILD_DIR/deb/$PKG_NAME-${PKG_VERSION}-${BIGTOP_BUILD_STAMP}"
+    def final ENABLE_MAVEN_PARALLEL_BUILD = config.bigtop.components[target].maven_parallel_build
+    def final MAVEN_BUILD_THREADS = project.hasProperty('buildThreads') ? project.property('buildThreads') : null
     mkdir (DEB_BLD_DIR)
     copy {
       from SEED_TAR
@@ -427,6 +451,12 @@ def genTasks = { target ->
         include '**/*'
       }
       into "$DEB_BLD_DIR/debian"
+    }
+    if (MAVEN_BUILD_THREADS && ENABLE_MAVEN_PARALLEL_BUILD) {
+      if (!MAVEN_BUILD_THREADS.matches("\\d+C")) {
+        throw new GradleException("Invalid MAVEN_BUILD_THREADS parameter. It must be a combination of numbers and 'C', such as '2C'.")
+      }
+      generatePatch(MAVEN_BUILD_THREADS, "$DEB_BLD_DIR/debian")
     }
     // Prepeare bom file with all the versions
     def bomWriter = new File("$DEB_BLD_DIR/debian/bigtop.bom").newWriter()
@@ -575,6 +605,8 @@ def genTasks = { target ->
     def final PKG_VERSION = config.bigtop.components[target].version.pkg
     def final BASE_VERSION = config.bigtop.components[target].version.base
     def final PKG_OUTPUT_DIR = config.bigtop.components[target].outputdir
+    def final ENABLE_MAVEN_PARALLEL_BUILD = config.bigtop.components[target].maven_parallel_build
+    def final MAVEN_BUILD_THREADS = project.hasProperty('buildThreads') ? project.property('buildThreads') : null
     safeDelete ("$PKG_BUILD_DIR/rpm")
     ['INSTALL','SOURCES','BUILD','SRPMS','RPMS'].each { rpmdir ->
       mkdir("$PKG_BUILD_DIR/rpm/$rpmdir")
@@ -608,6 +640,12 @@ def genTasks = { target ->
         include '**/*'
       }
       into "$PKG_BUILD_DIR/rpm/SOURCES"
+    }
+    if (MAVEN_BUILD_THREADS && ENABLE_MAVEN_PARALLEL_BUILD) {
+      if (!MAVEN_BUILD_THREADS.matches("\\d+C")) {
+        throw new GradleException("Invalid MAVEN_BUILD_THREADS parameter. It must be a combination of numbers and 'C', such as '2C'.")
+      }
+      generatePatch(MAVEN_BUILD_THREADS, "$PKG_BUILD_DIR/rpm/SOURCES")
     }
     // Writing bigtop.bom files with all the versions
     def bomWriter = new File("$PKG_BUILD_DIR/rpm/SOURCES/bigtop.bom").newWriter()