@JiaLiangC 2023-12-19
> What machine you uesd？
@MacChen01 
centos7 16C 16G SSD and maven dependency already downloaded.