@iwasakims 2023-12-20
@JiaLiangC  I got following error on Ubuntu 22.04 x86_64.

```                                                                                                                                                                                                                                                                                                         
$ ./gradlew zookeeper-clean zookeeper-pkg                                                                                                                                                                                                                                                                   
...                                                                                                                                                                                                                                                                                                         
+ mvn clean install -DskipTests -Pfull-build                                                                                                                                                                                                                                                                
[ERROR] Error executing Maven.                                                                                                                                                                                                                                                                              
[ERROR] Invalid threads value: '@CORES@'. Supported are int values.                                                                                                                                                                                                                                         
make[1]: *** [debian/rules:32: override_dh_auto_build] Error 1                                                                                                                                                                                                                                              
make[1]: Leaving directory '/home/iwasakims/srcs/bigtop/output/zookeeper/zookeeper-3.7.2'                                                                                                                                                                                                                   
make: *** [debian/rules:29: build] Error 2                                                                                                                                                                                                                                                                  
dpkg-buildpackage: error: debian/rules build subprocess returned exit status 2                                                                                                                                                                                                                              
```