@iwasakims 2023-12-20
How about conditionally adding Maven argument as [done for some products in do-component-build](https://github.com/apache/bigtop/blob/04830bd46ac4bf3cd607760b91be65a73c6579ba/bigtop-packages/src/common/hbase/do-component-build#L23-L55) instead of using `.mvn/maven.config`? We can access vars in bigtop.bom via environment variables.