@JiaLiangC 2023-12-20
@iwasakims 
Good idea! We just need to pass in parameters externally, and then, if the component supports concurrent compilation, we can simply handle the passed-in parameters by adding the corresponding Maven options during the do-component-build process. For components that do not support this, there's no need to process the parameter! I'll make some modifications and then test it again. 

Additionally, I will provide a comparison of the time consumption before and after enabling concurrent compilation for all components.