@JiaLiangC 2023-12-20
> How about conditionally adding Maven argument as [done for some products in do-component-build](https://github.com/apache/bigtop/blob/04830bd46ac4bf3cd607760b91be65a73c6579ba/bigtop-packages/src/common/hbase/do-component-build#L23-L55) instead of using `.mvn/maven.config`? We can access vars in bigtop.bom via environment variables.

@iwasakims 
I've reconsidered your approach and there's something I don't quite understand. What you're suggesting is:

The user compiles with a parallel compilation parameter, like compileThreads.
This parameter is then defined as an environmental variable.
Next, in the spec file or control file, this variable is passed to the do-component script for processing.
What's the benefit of this approach? It seems like it would lead to extensive, invasive code changes, requiring modifications to the do-component scripts and either the spec or control scripts of all involved components.