@JiaLiangC 2023-12-20
I have another idea:
We could save the mvn patch file in the template directory. For components that support concurrent compilation, we can simply copy it over. To determine which components support concurrent compilation, we might need to add an attribute in the bigtop.bom file to mark those components that support this feature.