@JiaLiangC 2023-12-25
@iwasakims Hello, based on our previous discussion about the method of introducing the parallel compilation patch, can we decide on a suitable plan? Then, I will proceed with subsequent testing based on this plan.