@iwasakims 2023-12-25
@JiaLiangC 
> We could save the mvn patch file in the template directory. For components that support concurrent compilation, we can simply copy it over.

It would be nice if this can cover both rpm and deb. I prefer directly generating the maven.config patch (in package.gradle from string template) to modifiying a file with regex.


> To determine which components support concurrent compilation, we might need to add an attribute in the bigtop.bom file to mark those components that support this feature.

I think just using `project.hasProperty('compileThreads')` as done in current patch would be enough with comments (and commented out defaults) in bigtop.bom.