@JiaLiangC 2023-12-26
@iwasakims 
Hello, I would like to ask about your reply mentioning "comments (and commented out defaults) in bigtop.bom". Does this refer to the 'label' under components, similar to how packaging = 'rpm' is used?

Currently, all Java components support concurrent compilation, so we only need to add maven_parallel_compile = False for the few components that do not support Java compilation. Alternatively, we could add maven_parallel_compile = True for each Java component individually. Which approach do you think is more appropriate?