@iwasakims 2023-12-26
@JiaLiangC How about `maven_parallel_build = True` prefering opt-in style and following [the name of the feature](https://cwiki.apache.org/confluence/display/MAVEN/Parallel+builds+in+Maven+3)? Parallel processing seems not to be limited to compilation based on your experience.