@JiaLiangC 2023-12-26
@iwasakims  If you mean just simply changing the label name from maven_parallel_compile to maven_parallel_build, in the style of https://cwiki.apache.org/confluence/display/MAVEN/Parallel+builds+in+Maven+3, then of course that's better.