@JiaLiangC 2023-12-28
@iwasakims 
Can you help review this PR? It has already been tested on CentOS 7 and Ubuntu 22.
 
![image](https://github.com/apache/bigtop/assets/18082602/9eb8c864-7304-4f79-8afc-a5aaa431f32a)